+++ src/QueryManager.ts
@@ -39,7 +39,6 @@ import {
   GraphQLResult,
   Document,
   FragmentDefinition,
-  OperationDefinition,
 } from 'graphql';
 
 import { print } from 'graphql-tag/printer';
@@ -294,21 +293,27 @@ export class QueryManager {
             console.error('Unhandled error', apolloError, apolloError.stack);
           }
         } else {
-          const resultFromStore = {
-            data: readSelectionSetFromStore({
-              store: this.getDataWithOptimisticResults(),
-              rootId: queryStoreValue.query.id,
-              selectionSet: queryStoreValue.query.selectionSet,
-              variables: queryStoreValue.variables,
-              returnPartialData: options.returnPartialData || options.noFetch,
-              fragmentMap: queryStoreValue.fragmentMap,
-            }),
-          };
+          try {
+            const resultFromStore = {
+              data: readSelectionSetFromStore({
+                store: this.getDataWithOptimisticResults(),
+                rootId: queryStoreValue.query.id,
+                selectionSet: queryStoreValue.query.selectionSet,
+                variables: queryStoreValue.variables,
+                returnPartialData: options.returnPartialData || options.noFetch,
+                fragmentMap: queryStoreValue.fragmentMap,
+              }),
+            };
 
-          if (observer.next) {
-            if (this.isDifferentResult(queryId, resultFromStore )) {
-              this.queryResults[queryId] = resultFromStore;
-              observer.next(resultFromStore);
+            if (observer.next) {
+              if (this.isDifferentResult(queryId, resultFromStore )) {
+                this.queryResults[queryId] = resultFromStore;
+                observer.next(resultFromStore);
+              }
+            }
+          } catch (error) {
+            if (observer.error) {
+              observer.error(error);
             }
           }
         }
@@ -510,7 +515,25 @@ export class QueryManager {
 
       queries.forEach((observableQuery) => {
         const queryOptions = observableQuery.options;
-        const queryDefinition: OperationDefinition = getQueryDefinition(queryOptions.query);
+
+        let fragments = queryOptions.fragments;
+        let queryDefinition = getQueryDefinition(queryOptions.query);
+
+        if (this.queryTransformer) {
+          const doc = {
+            kind: 'Document',
+            definitions: [
+              queryDefinition,
+              ...(fragments || []),
+            ],
+          };
+
+          const transformedDoc = applyTransformers(doc, [this.queryTransformer]);
+
+          queryDefinition = getQueryDefinition(transformedDoc);
+          fragments = getFragmentDefinitions(transformedDoc);
+        }
+
         const previousResult = readSelectionSetFromStore({
           // In case of an optimistic change, apply reducer on top of the
           // results including previous optimistic updates. Otherwise, apply it
@@ -520,7 +543,7 @@ export class QueryManager {
           selectionSet: queryDefinition.selectionSet,
           variables: queryOptions.variables,
           returnPartialData: queryOptions.returnPartialData || queryOptions.noFetch,
-          fragmentMap: createFragmentMap(queryOptions.fragments || []),
+          fragmentMap: createFragmentMap(fragments || []),
         });
 
         resultBehaviors.push({
@@ -530,7 +553,9 @@ export class QueryManager {
             queryName,
             queryVariables: queryOptions.variables,
           }),
-          queryOptions,
+          queryVariables: queryOptions.variables,
+          querySelectionSet: queryDefinition.selectionSet,
+          queryFragments: fragments,
         });
       });
     });