+++ test/client.ts
@@ -358,7 +358,11 @@ describe('client', () => {
             people: [ 'ROOT_QUERY.allPeople({"first":"1"}).people.0' ],
           },
           ROOT_QUERY: {
-            'allPeople({"first":1})': 'ROOT_QUERY.allPeople({"first":1})',
+            'allPeople({"first":1})': {
+              type: 'id',
+              id: 'ROOT_QUERY.allPeople({"first":1})',
+              generated: true,
+            },
           },
         },
         optimistic: [],
@@ -1093,6 +1097,29 @@ describe('client', () => {
       const expFragmentDefs = getFragmentDefinitions(otherFragmentDoc)
         .concat(getFragmentDefinitions(fragmentDoc));
       assert.deepEqual(fragmentDefs.map(print), expFragmentDefs.map(print));
+    });
+
+    it('should always return a flat array of fragment defs', () => {
+      const fragmentDoc1 = gql`
+        fragment authorDetails on Author {
+          firstName
+          lastName
+          ...otherAuthorDetails
+        }`;
+      const fragmentDoc2 = gql`
+        fragment otherAuthorDetails on Author {
+          address
+        }`;
+      const fragmentDoc3 = gql`
+        fragment personDetails on Person {
+          personDetails
+        }`;
+      const fragments1 = createFragment(fragmentDoc1);
+      const fragments2 = createFragment(fragmentDoc2, fragments1);
+      const fragments3 = createFragment(fragmentDoc3, fragments2);
+      assert.equal(fragments1.length, 1);
+      assert.equal(fragments2.length, 2);
+      assert.equal(fragments3.length, 3);
     });
 
     it('should add a fragment to the fragmentDefinitionsMap', () => {